requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
python-dotenv==1.0.1
PyYAML==6.0.1
openai==1.25.1
httpx==0.27.0
pytest==8.2.1
python-multipart